Move auto stuff to a place that makes sense
