+++ src/main/java/frc/robot/Robot.java
@@ -11,8 +11,8 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import frc.robot.Config.Key;
 import frc.robot.OI.OperatorInterface;
 import frc.robot.subsystems.*;
-import frc.robot.subsystems.auto.path.Path;
-import frc.robot.subsystems.auto.path.Paths;
+import frc.robot.auto.Path;
+import frc.robot.auto.Paths;
 import frc.robot.util.LatchedBoolean;
 import frc.robot.util.loops.Looper;
 import frc.robot.vision.AimingParameters;
