+++ src/main/java/frc/robot/auto/ShootSegment.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.auto.path;
+package frc.robot.auto;
 
 import frc.robot.OI.OperatorInterface;
 import frc.robot.subsystems.Storage;
